requirements.txt
Brotli
certifi
mutagen
pycryptodomex
pyperclip
websockets
PyYAML